Refactor Onboarding Wizard into smaller components and add isSubmitting state to prevent double clicks
- Extracted monolithic `page.tsx` into step-specific components under `components/`.
- Introduced `isSubmitting` flag in `page.tsx` to handle loading states for critical actions (`handleIntake`, `handleSendChatMessage`, `handleStartOnboarding`, `handleInstantBuild`).
- Updated E2E and Unit test suite (`page.test.tsx`, `e2e/onboarding.spec.ts`) to match new component structures and validations.
- Ensured premium macOS styling standards across components.
