+++ src/ui/next/src/app/onboarding/page.tsx
@@ -5,15 +5,14 @@ import { useRouter } from 'next/navigation';
 import { useOnboardingStore } from './store';
 import { SetupIcon } from './components/SetupIcon';
 import { IconLabel } from './components/IconLabel';
-
-function generateSubdomain(name: string): string {
-  if (!name || name.trim() === '') return 'my-business.ohc.app';
-  const cleanName = name
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '');
-  return cleanName ? `${cleanName}.ohc.app` : 'my-business.ohc.app';
-}
+import { StepIntro } from './components/StepIntro';
+import { StepChat } from './components/StepChat';
+import { StepInstantBuild } from './components/StepInstantBuild';
+import { StepChatQuestion } from './components/StepChatQuestion';
+import { StepReview } from './components/StepReview';
+import { StepStyleTeam } from './components/StepStyleTeam';
+import { StepLoading } from './components/StepLoading';
+import { StepSuccess } from './components/StepSuccess';
 
 export default function OnboardingWizard() {
   const router = useRouter();
@@ -31,11 +30,7 @@ export default function OnboardingWizard() {
   const [chatMessages, setChatMessages] = useState<{role: string, content: string, image_url?: string}[]>([]);
   const [chatInput, setChatInput] = useState('');
   const [chatImageUrl, setChatImageUrl] = useState('');
-  const chatMessagesEndRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    chatMessagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-  }, [chatMessages]);
+  const [isSubmitting, setIsSubmitting] = useState(false);
 
   const fetchWithRetry = async (url: string, options: RequestInit, retries = 3, backoff = process.env.NODE_ENV === 'test' ? 10 : 500) => {
     for (let i = 0; i < retries; i++) {
@@ -265,6 +260,8 @@ export default function OnboardingWizard() {
   ]);
 
   const handleIntake = async () => {
+    if (isSubmitting) return;
+    setIsSubmitting(true);
     updateState({ isLoading: true });
     updateState({ error: '' });
 
@@ -323,12 +320,14 @@ export default function OnboardingWizard() {
       updateState({ chatStep: 3 }); syncStateToBackend({ chatStep: 3 });
     } finally {
       updateState({ isLoading: false });
+      setIsSubmitting(false);
     }
   };
 
   const handleSendChatMessage = async () => {
-    if (!chatInput.trim() && !chatImageUrl.trim()) return;
+    if ((!chatInput.trim() && !chatImageUrl.trim()) || isSubmitting) return;
 
+    setIsSubmitting(true);
     const newMessage = {
       role: 'user',
       content: chatInput,
@@ -432,14 +431,16 @@ export default function OnboardingWizard() {
       updateState({ error: err.message || 'Failed to send chat message' });
     } finally {
       updateState({ isLoading: false });
+      setIsSubmitting(false);
     }
   };
 
   const handleInstantBuild = async () => {
-    if (!bio.trim()) {
-      updateState({ error: 'Please tell us about your business.' });
+    if (!bio.trim() || isSubmitting) {
+      if (!bio.trim()) updateState({ error: 'Please tell us about your business.' });
       return;
     }
+    setIsSubmitting(true);
     updateState({ isLoading: true });
     updateState({ error: '' });
 
@@ -494,10 +495,13 @@ export default function OnboardingWizard() {
       updateState({ step: -1 }); syncStateToBackend({ step: -1 });
     } finally {
       updateState({ isLoading: false });
+      setIsSubmitting(false);
     }
   };
 
   const handleStartOnboarding = async () => {
+    if (isSubmitting) return;
+
     const errors: Record<string, string> = {};
     if (!adminName.trim()) {
       errors.adminName = 'Admin Name is required';
@@ -517,6 +521,7 @@ export default function OnboardingWizard() {
       return;
     }
     setValidationErrors({});
+    setIsSubmitting(true);
     updateState({ isLoading: true });
     updateState({ error: '' });
     updateState({ step: 4 }); syncStateToBackend({ step: 4 }); // Go to loading screen
@@ -577,6 +582,7 @@ export default function OnboardingWizard() {
       updateState({ step: 3 }); syncStateToBackend({ step: 3 });
     } finally {
       updateState({ isLoading: false });
+      setIsSubmitting(false);
     }
   };
 
@@ -638,849 +644,53 @@ export default function OnboardingWizard() {
         )}
 
         <div className="p-6 flex-1 flex flex-col overflow-y-auto custom-scrollbar relative">
-          {step === -2 && (
-            <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-              <div className="w-16 h-16 bg-[#eef2ff] dark:bg-[#0066FF]/20 rounded-full flex items-center justify-center mb-6">
-                <svg className="w-8 h-8 text-[#0066FF]" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
-                </svg>
-              </div>
-              <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">10-Minute Setup Wizard</h2>
-              <p className="text-gray-500 dark:text-[#A1A1A6] text-sm text-center mb-8 leading-relaxed max-w-sm">
-                Zero tech skills needed. We do the heavy lifting. Review and add any extra details to help our AI generate the perfect store.
-              </p>
-
-              <div className="flex flex-col gap-4 w-full">
-                <button
-                  className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#005bb5] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] rounded-[8px]"
-                  onClick={() => { updateState({ step: 1 }); syncStateToBackend({ step: 1 }); }}
-                >
-                  Start My Business
-                </button>
-                <button
-                  type="button"
-                  className="w-full bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] p-4 font-semibold hover:border-gray-400 dark:hover:border-gray-500 transition-all"
-                  onClick={() => { updateState({ bio: "" }); updateState({ step: -1 }); syncStateToBackend({ step: -1, bio: "" }); }}
-                >
-                  Instant Build
-                </button>
-                <button
-                  type="button"
-                  className="w-full bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] p-4 font-semibold hover:border-gray-400 dark:hover:border-gray-500 transition-all"
-                  onClick={() => { updateState({ step: -2 }); syncStateToBackend({ step: -2 }); }}
-                >
-                  Conversational Setup
-                </button>
-              </div>
-            </div>
-          )}
-
-          {step === 0 && (
-            <div className="flex flex-col flex-1 animate-fade-in w-full h-full max-h-full">
-              <button onClick={() => { updateState({ step: -2 }); syncStateToBackend({ step: -2 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-              </button>
-              <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2 text-center">Setup Assistant</h2>
-              <p className="text-gray-500 dark:text-[#A1A1A6] text-sm text-center mb-4 leading-relaxed max-w-sm mx-auto">
-                Talk to our AI to build your business.
-              </p>
-
-              <div className="flex flex-col flex-1 gap-4 overflow-hidden w-full max-w-full">
-                <div id="chat-messages" className="bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[16px] flex-1 overflow-y-auto p-4 text-[#1D1D1F] dark:text-[#F5F5F7] text-left space-y-4">
-                  {chatMessages.length === 0 && (
-                    <div className="mb-2"><strong>Assistant:</strong> What do you do? (e.g. I bake custom vegan cakes in Austin)</div>
-                  )}
-                  {chatMessages.map((msg, index) => (
-                    <div key={index} className={`mb-2 ${msg.role === 'user' ? 'text-[#0066FF]' : 'text-[#333] dark:text-[#A1A1A6]'}`}>
-                      <strong>{msg.role === 'user' ? 'You' : 'Assistant'}:</strong> {msg.content}
-                      {msg.image_url && <><br /><span className="text-xs text-gray-500 dark:text-gray-400">[Attached Image: {msg.image_url}]</span></>}
-                    </div>
-                  ))}
-                  {isLoading && (
-                    <div className="mb-2 text-[#333] dark:text-[#A1A1A6]">
-                       <span className="flex items-center gap-2">
-                        <svg className="animate-spin h-4 w-4" fill="none" viewBox="0 0 24 24">
-                          <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                          <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                        </svg>
-                        <strong>Assistant:</strong> Thinking...
-                      </span>
-                    </div>
-                  )}
-                </div>
-
-                <div className="flex flex-col gap-2 shrink-0">
-                  <input
-                    type="url"
-                    id="chat-image-url"
-                    value={chatImageUrl}
-                    onChange={(e) => setChatImageUrl(e.target.value)}
-                    className="bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] w-full p-3 text-[#1D1D1F] dark:text-[#F5F5F7] outline-none transition-all duration-[250ms] border border-white/20 focus:border-[#0066FF] min-h-[44px]"
-                    placeholder="Image URL (Optional)"
-                    inputMode="url"
-                    autoComplete="url"
-                    enterKeyHint="next"
-                  />
-                  <div className="flex gap-2 w-full">
-                    <button
-                      id="chat-upload-btn"
-                      className="bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] min-w-[44px] min-h-[44px] flex items-center justify-center text-[#1D1D1F] dark:text-[#F5F5F7] hover:border-gray-400 dark:hover:border-gray-500 transition-all duration-[250ms] active:scale-[0.98]"
-                      onClick={() => {
-                        const url = prompt("Enter image URL");
-                        if (url) setChatImageUrl(url);
-                      }}
-                      title="Upload Image"
-                      aria-label="Upload Image"
-                    >
-                      <svg width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-                        <rect x="3" y="3" width="18" height="18" rx="2" ry="2"></rect>
-                        <circle cx="8.5" cy="8.5" r="1.5"></circle>
-                        <polyline points="21 15 16 10 5 21"></polyline>
-                      </svg>
-                    </button>
-                    <input
-                      type="text"
-                      id="chat-input"
-                      value={chatInput}
-                      onChange={(e) => setChatInput(e.target.value)}
-                      onKeyDown={(e) => {
-                         if (e.key === 'Enter') handleSendChatMessage();
-                      }}
-                      className="bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] w-full p-3 text-[#1D1D1F] dark:text-[#F5F5F7] outline-none flex-1 transition-all duration-[250ms] border border-white/20 focus:border-[#0066FF] min-h-[44px]"
-                      placeholder="Type a message..."
-                      enterKeyHint="send"
-                    />
-                    <button
-                      id="chat-send-btn"
-                      onClick={handleSendChatMessage}
-                      disabled={isLoading}
-                      className="bg-[#0066FF] text-white font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#005bb5] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] px-4 shrink-0 disabled:opacity-50 rounded-[8px]"
-                    >
-                      Send
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
-          )}
-
-          {step === -1 && (
-            <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-              <button onClick={() => { updateState({ step: -2 }); syncStateToBackend({ step: -2 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-              </button>
-              <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">Tell us about your business</h2>
-              <p className="text-gray-500 dark:text-[#A1A1A6] text-sm text-center mb-8 leading-relaxed max-w-sm">
-                Our AI will handle the rest in 30 seconds.
-              </p>
-
-              <div className="flex flex-col gap-4 w-full">
-                <textarea
-                  id="instant-bio"
-                  data-testid="instant-bio"
-                  className={`bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] w-full p-4 text-[#1D1D1F] dark:text-[#F5F5F7] outline-none transition-all duration-[250ms] ${error === "Please tell us about your business." || error ? "border border-[#FF3B30]" : "border border-white/20 focus:border-[#0066FF]"}`}
-                  placeholder="e.g. I run a local bakery that sells custom vegan cakes..."
-                  rows={6}
-                  style={{ resize: 'none' }}
-                  value={bio}
-                  onChange={(e) => {
-                    updateState({ bio: e.target.value });
-                    if (error) updateState({ error: '' });
-                  }}
-                />
-
-                <input
-                  id="instant-image-url"
-                  data-testid="instant-image-url"
-                  type="url"
-                  className="bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] min-h-[44px]"
-                  placeholder="Image URL (Optional)"
-                  value={instantImageUrl}
-                  onChange={(e) => updateState({ instantImageUrl: e.target.value })}
-                  inputMode="url"
-                  autoComplete="url"
-                />
-
-                <div className="mt-4">
-                  <button
-                    onClick={handleInstantBuild}
-                    disabled={!bio.trim() || isLoading}
-                    className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#005bb5] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed rounded-[8px]"
-                  >
-                    Next
-                  </button>
-                </div>
-              </div>
-            </div>
-          )}
-
+          {step === -2 && <StepIntro syncStateToBackend={syncStateToBackend} />}
+          {step === -1 && <StepInstantBuild handleInstantBuild={handleInstantBuild} syncStateToBackend={syncStateToBackend} />}
+          {step === 0 && <StepChat
+            chatMessages={chatMessages}
+            setChatMessages={setChatMessages}
+            chatInput={chatInput}
+            setChatInput={setChatInput}
+            chatImageUrl={chatImageUrl}
+            setChatImageUrl={setChatImageUrl}
+            handleSendChatMessage={handleSendChatMessage}
+            syncStateToBackend={syncStateToBackend}
+          />}
           {step === 1 && (
-            <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-              <div className="w-16 h-16 bg-[#eef2ff] dark:bg-[#0066FF]/20 rounded-full flex items-center justify-center mb-6">
-                <svg className="w-8 h-8 text-[#0066FF]" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
-                </svg>
-              </div>
-
-              {chatStep === 1 && (
-                <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-                  <button onClick={() => { updateState({ step: -2 }); syncStateToBackend({ step: -2 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1">
-                    <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-                  </button>
-                  <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">What's the name of your business?</h2>
-                  <div className="flex items-start sm:items-center justify-between mb-6 w-full gap-2">
-                    <p className="text-gray-500 dark:text-[#A1A1A6] text-sm pr-4">
-                      Our AI will instantly generate your storefront, products, and back-office agents.
-                    </p>
-                    <button
-                      onClick={() => handleSaveDraft()}
-                      className="text-sm font-semibold text-[#0066FF] hover:underline whitespace-nowrap shrink-0 ml-auto flex items-center justify-center min-h-[44px] min-w-[44px] p-2"
-                    >
-                      <IconLabel icon="save">Save Draft</IconLabel>
-                    </button>
-                  </div>
-
-                  {saveMessage && <p className="text-[#34C759] text-sm font-semibold mb-2">{saveMessage}</p>}
-
-                  <div className="space-y-4 flex-1">
-                    <div>
-                      <input
-                        type="text"
-                        autoFocus
-
-                        autoCapitalize="words"
-                        autoComplete="organization"
-                        value={businessName}
-                        onChange={(e) => {
-                          const val = e.target.value;
-                          updateState({ businessName: val });
-                          if (val.trim().length < 3) { setValidationError('Business Name must be at least 3 characters.'); }
-                          else { setValidationError(''); }
-                        }}
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter') {
-                            e.preventDefault();
-                            if (businessName.trim().length < 3) {
-                              setValidationError('Business Name must be at least 3 characters.');
-                              return;
-                            }
-                            setValidationError('');
-                            updateState({ chatStep: 2 }); syncStateToBackend({ chatStep: 2 });
-                          }
-                        }}
-                        placeholder="e.g. Maya's Custom Cakes"
-                        className={`w-full p-3 sm:p-4 border outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] text-lg transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] shadow-inner ${validationError === 'Business Name must be at least 3 characters.' ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]'} min-h-[44px]`}
-                        inputMode="text"
-                        enterKeyHint="next"
-                      />
-                    </div>
-                  </div>
-
-                  {validationError && <p className="text-[#FF3B30] text-sm font-semibold mb-2">{validationError}</p>}
-                  <div className="mt-auto pt-6">
-                    <button
-                      onClick={() => {
-                        if (businessName.trim().length < 3) {
-                          setValidationError('Business Name must be at least 3 characters.');
-                          return;
-                        }
-                        setValidationError('');
-                        updateState({ chatStep: 2 }); syncStateToBackend({ chatStep: 2 });
-                      }}
-                      disabled={false}
-                      className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#0052cc] hover:shadow-[0_6px_20px_rgba(0,102,255,0.23)] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      <IconLabel icon="next">Next</IconLabel>
-                    </button>
-                  </div>
-                </div>
-              )}
-
-              {chatStep === 2 && (
-                <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-                  <button onClick={() => { updateState({ chatStep: 1 }); syncStateToBackend({ chatStep: 1 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                    <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-                  </button>
-                  <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">What do you sell?</h2>
-                  <div className="flex items-start sm:items-center justify-between mb-6 w-full gap-2">
-                    <p className="text-gray-500 dark:text-[#A1A1A6] text-sm pr-4">
-                      Tell us a bit about your products or services.
-                    </p>
-                    <button
-                      onClick={() => handleSaveDraft()}
-                      className="text-sm font-semibold text-[#0066FF] hover:underline whitespace-nowrap shrink-0 ml-auto flex items-center justify-center min-h-[44px] min-w-[44px] p-2"
-                    >
-                      <IconLabel icon="save">Save Draft</IconLabel>
-                    </button>
-                  </div>
-
-                  {saveMessage && <p className="text-[#34C759] text-sm font-semibold mb-2">{saveMessage}</p>}
-
-                  <div className="space-y-4 flex-1">
-                    <div>
-                      <textarea
-                        autoFocus
-
-                        autoCapitalize="sentences"
-                        value={whatYouSell}
-                        onChange={(e) => {
-                          const val = e.target.value;
-                          updateState({ whatYouSell: val });
-                          if (!val.trim()) { setValidationError('Please tell us what you sell.'); }
-                          else { setValidationError(''); }
-                        }}
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter' && !e.shiftKey) {
-                            e.preventDefault();
-                            if (!whatYouSell.trim()) {
-                              setValidationError('Please tell us what you sell.');
-                              return;
-                            }
-                            setValidationError('');
-                            updateState({ chatStep: 3 }); syncStateToBackend({ chatStep: 3 });
-                          }
-                        }}
-                        placeholder="e.g. I bake custom vegan cakes"
-                        className={`w-full p-3 sm:p-4 border outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] h-32 resize-none transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] shadow-inner ${validationError === 'Please tell us what you sell.' ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF] focus:ring-2 focus:ring-[#0066FF]/30'}`}
-                      />
-                    </div>
-                  </div>
-
-                  {validationError && <p className="text-[#FF3B30] text-sm font-semibold mb-2">{validationError}</p>}
-                  <div className="mt-auto pt-6">
-                    <button
-                      onClick={() => {
-                        if (!whatYouSell.trim()) {
-                          setValidationError('Please tell us what you sell.');
-                          return;
-                        }
-                        setValidationError('');
-                        updateState({ chatStep: 3 }); syncStateToBackend({ chatStep: 3 });
-                      }}
-                      disabled={false}
-                      className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#0052cc] hover:shadow-[0_6px_20px_rgba(0,102,255,0.23)] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      <IconLabel icon="next">Next</IconLabel>
-                    </button>
-                  </div>
-                </div>
-              )}
-
-              {chatStep === 3 && (
-                <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-                  <button onClick={() => { updateState({ chatStep: 2 }); syncStateToBackend({ chatStep: 2 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                    <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-                  </button>
-                  <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">Where are you located?</h2>
-                  <div className="flex items-start sm:items-center justify-between mb-6 w-full gap-2">
-                    <p className="text-gray-500 dark:text-[#A1A1A6] text-sm pr-4">
-                      This helps us set up your shipping and tax settings.
-                    </p>
-                    <button
-                      onClick={() => handleSaveDraft()}
-                      className="text-sm font-semibold text-[#0066FF] hover:underline whitespace-nowrap shrink-0 ml-auto flex items-center justify-center min-h-[44px] min-w-[44px] p-2"
-                    >
-                      <IconLabel icon="save">Save Draft</IconLabel>
-                    </button>
-                  </div>
-
-                  {saveMessage && <p className="text-[#34C759] text-sm font-semibold mb-2">{saveMessage}</p>}
-
-                  <div className="space-y-4 flex-1">
-                    <div>
-                      <input
-                        type="text"
-                        autoFocus
-
-                        autoCapitalize="words"
-                        value={location}
-                        onChange={(e) => updateState({ location: e.target.value })}
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter') {
-                            e.preventDefault();
-                            e.stopPropagation();
-                            if (!location.trim()) {
-                              setValidationError('Please tell us your location.');
-                              return;
-                            }
-                            setValidationError('');
-                            updateState({ chatStep: 4 }); syncStateToBackend({ chatStep: 4 });
-                          }
-                        }}
-                        placeholder="e.g. Portland, OR"
-                        className={`w-full p-3 sm:p-4 border outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] text-lg transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] shadow-inner ${validationError === 'Please tell us your location.' ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]'} min-h-[44px]`}
-                      />
-                    </div>
-                  </div>
-
-                  {validationError && <p className="text-[#FF3B30] text-sm font-semibold mb-2">{validationError}</p>}
-                  <div className="mt-auto pt-6">
-                    <button
-                      onClick={() => {
-                        if (!location.trim()) {
-                          setValidationError('Please tell us your location.');
-                          return;
-                        }
-                        setValidationError('');
-                        updateState({ chatStep: 4 }); syncStateToBackend({ chatStep: 4 });
-                      }}
-                      disabled={false}
-                      className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#0052cc] hover:shadow-[0_6px_20px_rgba(0,102,255,0.23)] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      <IconLabel icon="next">Next</IconLabel>
-                    </button>
-                  </div>
-                </div>
-              )}
-
-              {chatStep === 4 && (
-                <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-                  <button onClick={() => { updateState({ chatStep: 3 }); syncStateToBackend({ chatStep: 3 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                    <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-                  </button>
-                  <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">Who is your target audience?</h2>
-                  <div className="flex items-start sm:items-center justify-between mb-6 w-full gap-2">
-                    <p className="text-gray-500 dark:text-[#A1A1A6] text-sm pr-4">
-                      This helps our AI generate the perfect storefront copy and select the best tools for your business.
-                    </p>
-                    <button
-                      onClick={() => handleSaveDraft()}
-                      className="text-sm font-semibold text-[#0066FF] hover:underline whitespace-nowrap shrink-0 ml-auto flex items-center justify-center min-h-[44px] min-w-[44px] p-2"
-                    >
-                      <IconLabel icon="save">Save Draft</IconLabel>
-                    </button>
-                  </div>
-
-                  {saveMessage && <p className="text-[#34C759] text-sm font-semibold mb-2">{saveMessage}</p>}
-
-                  <div className="space-y-4 flex-1">
-                    <div>
-                      <input
-                        type="text"
-                        autoFocus
-
-                        autoCapitalize="words"
-                        value={targetAudience}
-                        onChange={(e) => updateState({ targetAudience: e.target.value })}
-                        onKeyDown={(e) => {
-                          if (e.key === 'Enter') {
-                            e.preventDefault();
-                            e.stopPropagation();
-                            if (!targetAudience.trim()) {
-                              setValidationError('Please tell us your target audience.');
-                              return;
-                            }
-                            setValidationError('');
-                            handleIntake();
-                          }
-                        }}
-                        placeholder="e.g. Local families, Tech startups"
-                        className={`w-full p-3 sm:p-4 border outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] text-lg transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] shadow-inner ${validationError === 'Please tell us your target audience.' ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]'} min-h-[44px]`}
-                      />
-                    </div>
-                  </div>
-
-                  {validationError && <p className="text-[#FF3B30] text-sm font-semibold mb-2">{validationError}</p>}
-                  <div className="mt-auto pt-6">
-                    <button
-                      onClick={() => {
-                        if (!targetAudience.trim()) {
-                          setValidationError('Please tell us your target audience.');
-                          return;
-                        }
-                        setValidationError('');
-                        handleIntake();
-                      }}
-                      disabled={isLoading}
-                      className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#0052cc] hover:shadow-[0_6px_20px_rgba(0,102,255,0.23)] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed"
-                    >
-                      {isLoading ? (
-                        <span className="flex items-center justify-center gap-2">
-                          <svg className="animate-spin h-5 w-5 text-white rounded-full shadow-[0_0_10px_rgba(255,255,255,0.5)]" style={{ backdropFilter: 'blur(30px) saturate(210%)', WebkitBackdropFilter: 'blur(30px) saturate(210%)' }} fill="none" viewBox="0 0 24 24">
-                            <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                            <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                          </svg>
-                          Analyzing...
-                        </span>
-                      ) : <IconLabel icon="launch">Next</IconLabel>}
-                    </button>
-                  </div>
-                </div>
-              )}
-            </div>
+             <StepChatQuestion
+               handleSaveDraft={handleSaveDraft}
+               saveMessage={saveMessage}
+               syncStateToBackend={syncStateToBackend}
+               validationError={validationError}
+               setValidationError={setValidationError}
+               handleIntake={handleIntake}
+               isSubmitting={isSubmitting}
+             />
           )}
-
           {step === 2 && (
-            <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-              <button onClick={() => { updateState({ step: 1 }); updateState({ chatStep: 4 }); syncStateToBackend({ step: 1, chatStep: 4 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-              </button>
-              <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">Review Details</h2>
-              <div className="flex items-start sm:items-center justify-between mb-6 w-full gap-2">
-                <p className="text-gray-500 dark:text-[#A1A1A6] text-sm pr-4">
-                  Here's what our AI figured out. Feel free to tweak these.
-                </p>
-                <button
-                  onClick={() => handleSaveDraft()}
-                  className="text-sm font-semibold text-[#0066FF] hover:underline whitespace-nowrap shrink-0 ml-auto flex items-center justify-center min-h-[44px] min-w-[44px] p-2"
-                >
-                      <IconLabel icon="save">Save Draft</IconLabel>
-                    </button>
-              </div>
-
-              {saveMessage && <p className="text-[#34C759] text-sm font-semibold mb-2">{saveMessage}</p>}
-
-              <div className="space-y-4 flex-1 overflow-y-auto pr-2">
-                <div>
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-1">Business Name</label>
-                  <input
-                    type="text"
-                    autoFocus
-
-                    autoCapitalize="words"
-                    value={businessName}
-                    onChange={(e) => {
-                      updateState({ businessName: e.target.value });
-                      setValidationErrors(prev => { const { businessName, ...rest } = prev; return rest; });
-                    }}
-                    className={`w-full p-3 sm:p-4 border ${validationErrors.businessName ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]'} outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]`}
-                  />
-                  {validationErrors.businessName && <p className="text-[#FF3B30] text-xs mt-1">{validationErrors.businessName}</p>}
-                </div>
-                <div>
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-1">Business Type</label>
-                  <input
-                    type="text"
-
-                    autoCapitalize="words"
-                    value={businessType}
-                    onChange={(e) => {
-                      updateState({ businessType: e.target.value });
-                      setValidationErrors(prev => { const { businessType, ...rest } = prev; return rest; });
-                    }}
-                    className={`w-full p-3 sm:p-4 border ${validationErrors.businessType ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]'} outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]`}
-                  />
-                  {validationErrors.businessType && <p className="text-[#FF3B30] text-xs mt-1">{validationErrors.businessType}</p>}
-                </div>
-                <div>
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-1">Categories (Comma separated)</label>
-                  <input
-                    type="text"
-
-                    autoCapitalize="words"
-                    value={categories.join(', ')}
-                    onChange={(e) => updateState({ categories: e.target.value.split(',').map(c => c.trim()) })}
-                    className="w-full p-3 sm:p-4 border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF] outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]"
-                  />
-                </div>
-                <div className="grid grid-cols-2 gap-2">
-                   <div>
-                      <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-1">First Product</label>
-                      <input
-                        type="text"
-
-                        autoCapitalize="words"
-                        value={firstProductName}
-                        onChange={(e) => updateState({ firstProductName: e.target.value })}
-                        className="w-full p-3 sm:p-4 border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF] outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]"
-                      />
-                   </div>
-                   <div>
-                      <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-1">Price</label>
-                      <input
-                        type="text"
-                        inputMode="decimal"
-                        value={firstProductPrice}
-                        onChange={(e) => {
-                           updateState({ firstProductPrice: e.target.value });
-                           if (e.target.value.trim().length > 0 && !/^\d+(\.\d{1,2})?$/.test(e.target.value)) {
-                              setValidationErrors(prev => ({ ...prev, firstProductPrice: 'Invalid price.' }));
-                           } else {
-                              setValidationErrors(prev => { const { firstProductPrice, ...rest } = prev; return rest; });
-                           }
-                        }}
-                        className={`w-full p-3 sm:p-4 border ${validationErrors.firstProductPrice ? 'border-[#FF3B30]' : 'border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]'} outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]`}
-                      />
-                      {validationErrors.firstProductPrice && <p className="text-[#FF3B30] text-xs mt-1">{validationErrors.firstProductPrice}</p>}
-                   </div>
-                </div>
-              </div>
-
-              {validationError && <p className="text-[#FF3B30] text-sm font-semibold mb-2">{validationError}</p>}
-              <div className="mt-auto pt-6">
-                <button
-                  onClick={() => {
-                    let hasError = false;
-                    const newErrors: Record<string, string> = { ...validationErrors };
-                    if (businessName.trim().length < 3) {
-                      newErrors.businessName = 'Must be at least 3 characters.';
-                      hasError = true;
-                    }
-                    if (businessType.trim().length === 0) {
-                      newErrors.businessType = 'Business Type is required to configure your agents.';
-                      hasError = true;
-                    }
-                    if (firstProductPrice.trim().length === 0) {
-                      newErrors.firstProductPrice = 'A price is needed to set up your Stripe catalog.';
-                      hasError = true;
-                    }
-
-                    if (hasError || Object.keys(newErrors).length > 0) {
-                      setValidationErrors(newErrors);
-                      setValidationError('Please fix the errors before continuing.');
-                      return;
-                    }
-
-                    setValidationError('');
-                    updateState({ step: 3 }); syncStateToBackend({ step: 3 });
-                  }}
-                  className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#0052cc] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  <IconLabel icon="next">Continue</IconLabel>
-                </button>
-              </div>
-            </div>
+             <StepReview
+               handleSaveDraft={handleSaveDraft}
+               saveMessage={saveMessage}
+               syncStateToBackend={syncStateToBackend}
+               validationErrors={validationErrors}
+               setValidationErrors={setValidationErrors}
+               validationError={validationError}
+               setValidationError={setValidationError}
+             />
           )}
-
           {step === 3 && (
-            <div className="flex flex-col justify-center items-center gap-4 flex-1 animate-fade-in">
-              <button onClick={() => { updateState({ step: 2 }); syncStateToBackend({ step: 2 }); }} className="self-start text-[#0066FF] text-sm font-semibold mb-4 flex items-center gap-1 min-h-[44px] min-w-[44px] p-2">
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 19l-7-7 7-7" /></svg> Back
-              </button>
-              <h2 className="text-3xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">Style & Team</h2>
-              <div className="flex items-start sm:items-center justify-between mb-6 w-full gap-2">
-                <p className="text-gray-500 dark:text-[#A1A1A6] text-sm pr-4">
-                  Pick your storefront vibe. We'll automatically assign the best AI agents to manage it.
-                </p>
-                <button
-                  onClick={() => handleSaveDraft()}
-                  className="text-sm font-semibold text-[#0066FF] hover:underline whitespace-nowrap shrink-0 ml-auto flex items-center justify-center min-h-[44px] min-w-[44px] p-2"
-                >
-                      <IconLabel icon="save">Save Draft</IconLabel>
-                    </button>
-              </div>
-
-              {saveMessage && <p className="text-[#34C759] text-sm font-semibold mb-2">{saveMessage}</p>}
-
-              <div className="space-y-4 flex-1 overflow-y-auto pr-2 hide-scrollbar">
-                <div>
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-2">Website Template</label>
-                  <div className="grid grid-cols-2 gap-3">
-                    {['Modern', 'Minimal', 'Bold', 'Classic'].map(template => (
-                      <div
-                        key={template}
-                        onClick={() => updateState({ websiteTemplate: template })}
-                        className={`p-3 border cursor-pointer transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] ${websiteTemplate === template ? 'border-[#0066FF] bg-[#0066FF]/10 text-[#0066FF]' : 'border-white/50 dark:border-[rgba(255,255,255,0.1)] bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] hover:border-gray-400 dark:hover:border-gray-500 text-[#1D1D1F] dark:text-white'}`}
-                      >
-                        <div className="font-semibold text-sm">{template}</div>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-
-                <div className="pt-2 border-t border-white/50 dark:border-[rgba(255,255,255,0.1)]">
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-2">Web Address</label>
-                  <div className="grid grid-cols-2 gap-3 mb-2">
-                    <div
-                      onClick={() => updateState({ domainChoice: 'subdomain' })}
-                      className={`p-3 border cursor-pointer transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] flex flex-col items-center justify-center text-center ${domainChoice === 'subdomain' ? 'border-[#0066FF] bg-[#0066FF]/10 text-[#0066FF]' : 'border-white/50 dark:border-[rgba(255,255,255,0.1)] bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-white hover:border-gray-400 dark:hover:border-gray-500'}`}
-                    >
-                      <span className="font-semibold text-sm mb-1">Free Subdomain</span>
-                      <span className="text-[10px] opacity-70">your-name.ohc.app</span>
-                    </div>
-                    <div
-                      onClick={() => updateState({ domainChoice: 'custom' })}
-                      className={`p-3 border cursor-pointer transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] flex flex-col items-center justify-center text-center ${domainChoice === 'custom' ? 'border-[#0066FF] bg-[#0066FF]/10 text-[#0066FF]' : 'border-white/50 dark:border-[rgba(255,255,255,0.1)] bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-white hover:border-gray-400 dark:hover:border-gray-500'}`}
-                    >
-                      <span className="font-semibold text-sm mb-1">Custom Domain</span>
-                      <span className="text-[10px] opacity-70">your-name.com</span>
-                    </div>
-                  </div>
-                </div>
-
-                <div className="pt-2 border-t border-white/50 dark:border-[rgba(255,255,255,0.1)]">
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-2">Account Setup</label>
-                  <div className="space-y-3 mb-4">
-                    <div>
-                      <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 mb-1">Admin Name</label>
-                      <input
-                        type="text"
-
-                        autoCapitalize="words"
-                        autoComplete="name"
-                        value={adminName}
-                        onChange={(e) => {
-                          const val = e.target.value;
-                          updateState({ adminName: val });
-                          if (!val.trim()) {
-                            setValidationErrors(prev => ({ ...prev, adminName: 'Admin Name is required' }));
-                          } else {
-                            setValidationErrors(prev => { const { adminName, ...rest } = prev; return rest; });
-                          }
-                        }}
-                        placeholder="e.g. Maya Smith"
-                        className={`w-full p-3 sm:p-4 border ${validationErrors.adminName ? "border-[#FF3B30]" : "border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]"} outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]`}
-                        inputMode="text"
-                        enterKeyHint="next"
-                      />
-                      {validationErrors.adminName && <p className="text-[#FF3B30] text-xs mt-1">{validationErrors.adminName}</p>}
-                    </div>
-                    <div>
-                      <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 mb-1">Admin Email</label>
-                      <input
-                        type="email"
-                    autoCapitalize="none"
-                    autoComplete="email"
-                    inputMode="email"
-                    enterKeyHint="next"
-                    value={adminEmail}
-                    onChange={(e) => {
-                      const val = e.target.value;
-                      updateState({ adminEmail: val });
-                      if (!val.trim()) {
-                        setValidationErrors(prev => ({ ...prev, adminEmail: 'Admin Email is required' }));
-                      } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(val)) {
-                        setValidationErrors(prev => ({ ...prev, adminEmail: 'Please enter a valid email address' }));
-                      } else {
-                        setValidationErrors(prev => { const { adminEmail, ...rest } = prev; return rest; });
-                      }
-                    }}
-                    placeholder="you@example.com"
-                    className={`w-full p-3 sm:p-4 border ${validationErrors.adminEmail ? "border-[#FF3B30]" : "border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]"} outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]`}
-                  />
-                      {validationErrors.adminEmail && <p className="text-[#FF3B30] text-xs mt-1">{validationErrors.adminEmail}</p>}
-                    </div>
-                    <div>
-                      <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 mb-1">Admin Password</label>
-                      <input
-                        type="password"
-                        enterKeyHint="done"
-                        autoComplete="new-password"
-                        value={adminPassword}
-                        onChange={(e) => {
-                          const val = e.target.value;
-                          updateState({ adminPassword: val });
-                          if (!val.trim()) {
-                            setValidationErrors(prev => ({ ...prev, adminPassword: 'Password is required' }));
-                          } else if (val.length < 8 || !/\d/.test(val)) {
-                            setValidationErrors(prev => ({ ...prev, adminPassword: 'Password must be at least 8 characters and contain a number' }));
-                          } else {
-                            setValidationErrors(prev => { const { adminPassword, ...rest } = prev; return rest; });
-                          }
-                        }}
-                        placeholder="••••••••"
-                        className={`w-full p-3 sm:p-4 border ${validationErrors.adminPassword ? "border-[#FF3B30]" : "border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] focus:border-[#0066FF]"} outline-none bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] min-h-[44px]`}
-                      />
-                      {validationErrors.adminPassword && <p className="text-[#FF3B30] text-xs mt-1">{validationErrors.adminPassword}</p>}
-                    </div>
-                  </div>
-                </div>
-
-                <div className="pt-2 border-t border-white/50 dark:border-[rgba(255,255,255,0.1)]">
-                  <label className="block text-xs font-semibold text-gray-700 dark:text-gray-300 uppercase tracking-wide mb-2">Auto-Configured AI Departments</label>
-                  <p className="text-gray-500 dark:text-[#A1A1A6] text-xs mb-2">
-                    Here are the AI departments we've configured for you.
-                  </p>
-                  <div className="flex flex-col sm:flex-row flex-wrap gap-2 mt-2">
-                    {aiAgents.map(agent => (
-                      <div
-                        key={agent}
-                        className="px-3 py-1.5 rounded-[8px] border border-[#34C759] bg-[#34C759]/10 text-[#34C759] flex items-center gap-1.5 text-sm font-semibold transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)]"
-                      >
-                        <svg className="w-3.5 h-3.5" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={3} d="M5 13l4 4L19 7" /></svg>
-                        {agent}
-                      </div>
-                    ))}
-                  </div>
-                </div>
-
-                <div className="pt-2">
-                  <label className="flex items-center justify-between cursor-pointer p-3 bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-white">
-                    <span className="font-semibold text-sm">Allow AI to Auto-Respond</span>
-                    <input
-                      type="checkbox"
-                      className="sr-only"
-                      checked={aiAutoRespond}
-                      onChange={(e) => updateState({ aiAutoRespond: e.target.checked })}
-                    />
-                    <div className={`w-10 h-6 rounded-full transition-colors ${aiAutoRespond ? 'bg-[#34C759]' : 'bg-gray-300 dark:bg-gray-600'} relative`}>
-                       <div className={`w-4 h-4 rounded-full bg-white absolute top-1 transition-transform ${aiAutoRespond ? 'translate-x-5' : 'translate-x-1'}`}></div>
-                    </div>
-                  </label>
-                </div>
-              </div>
-
-              <div className="mt-auto pt-6">
-                <button
-                  onClick={() => handleStartOnboarding()}
-                  disabled={isLoading}
-                  className="w-full bg-[#0066FF] text-white p-4 font-bold shadow-[0_4px_14px_0_rgba(0,102,255,0.39)] hover:bg-[#0052cc] active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)] disabled:opacity-50 disabled:cursor-not-allowed"
-                >
-                  {isLoading ? (
-                    <span className="flex items-center justify-center gap-2">
-                      <svg className="animate-spin h-5 w-5 text-white" fill="none" viewBox="0 0 24 24">
-                        <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                        <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                      </svg>
-                      Launching...
-                    </span>
-                  ) : <IconLabel icon="launch">Approve & Publish</IconLabel>}
-                </button>
-              </div>
-            </div>
+             <StepStyleTeam
+               handleStartOnboarding={handleStartOnboarding}
+               handleSaveDraft={handleSaveDraft}
+               saveMessage={saveMessage}
+               syncStateToBackend={syncStateToBackend}
+               validationErrors={validationErrors}
+               setValidationErrors={setValidationErrors}
+               isSubmitting={isSubmitting}
+             />
           )}
-
-          {step === 4 && (
-             <div aria-live="polite" className="flex flex-col flex-1 justify-center items-center text-center animate-fade-in bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[16px] shadow-2xl p-4 sm:p-8">
-               <div className="w-24 h-24 relative mb-8">
-                 <div className="absolute inset-0 border-4 border-[#0066FF]/20 rounded-full"></div>
-                 <div className="absolute inset-0 border-4 border-[#0066FF] rounded-full border-t-transparent animate-spin"></div>
-               </div>
-               <h2 className="text-2xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-4">Building Your Business...</h2>
-               <div className="space-y-2">
-                 <p className="text-gray-500 dark:text-[#A1A1A6] text-sm animate-pulse">Generating your product catalog</p>
-                 <p className="text-gray-500 dark:text-[#A1A1A6] text-sm animate-pulse" style={{ animationDelay: '0.5s' }}>Configuring payment settings</p>
-                 <p className="text-gray-500 dark:text-[#A1A1A6] text-sm animate-pulse" style={{ animationDelay: '1s' }}>Designing your storefront</p>
-                 <p className="text-gray-500 dark:text-[#A1A1A6] text-sm animate-pulse" style={{ animationDelay: '1.5s' }}>Onboarding your AI agents</p>
-               </div>
-             </div>
-          )}
-
-          {step === 5 && startResult && (
-            <div className="flex flex-col flex-1 justify-center items-center text-center animate-fade-in">
-              <div className="w-20 h-20 bg-[#34C759]/20 rounded-full flex items-center justify-center mb-6">
-                <svg className="w-10 h-10 text-[#34C759]" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={3} d="M5 13l4 4L19 7" />
-                </svg>
-              </div>
-              <h2 className="text-2xl font-bold font-outfit text-[#1D1D1F] dark:text-[#F5F5F7] mb-2">You're Live!</h2>
-              <p className="text-gray-500 dark:text-[#A1A1A6] text-sm mb-8 px-4">
-                {startResult.message || "Your business has been successfully launched."}
-              </p>
-
-              <div className="w-full space-y-3 mt-auto">
-                <div className="p-3 bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[16px] flex flex-col items-center mb-6">
-                   <p className="text-xs text-gray-500 dark:text-[#A1A1A6] uppercase font-bold tracking-wider mb-2">Your Shareable Link</p>
-                   <div className="flex items-center gap-2">
-                      <span className="text-[#0066FF] font-semibold">{generateSubdomain(businessName)}</span>
-                   </div>
-                </div>
-
-                <a
-                  href="/assistant"
-                  className="flex w-full items-center justify-center bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] p-4 font-bold shadow-md hover:border-gray-400 dark:hover:border-gray-500 active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)]"
-                >
-                  <IconLabel icon="sparkles">Open Assistant</IconLabel>
-                </a>
-                <a
-                  href="/builder"
-                  className="flex w-full items-center justify-center bg-[rgba(255,255,255,0.65)] dark:bg-[rgba(22,22,26,0.7)] backdrop-blur-[30px] backdrop-saturate-[210%] border border-[rgba(255,255,255,0.4)] dark:border-[rgba(255,255,255,0.1)] rounded-[8px] text-[#1D1D1F] dark:text-[#F5F5F7] p-4 font-bold shadow-sm active:scale-[0.98] transition-all duration-[250ms] ease-[cubic-bezier(0.4,0,0.2,1)]"
-                >
-                  <IconLabel icon="eye">Preview Storefront</IconLabel>
-                </a>
-              </div>
-            </div>
-          )}
+          {step === 4 && <StepLoading />}
+          {step === 5 && startResult && <StepSuccess />}
         </div>
       </div>
     </div>
